onimaru animations fixed, death problems fixed, lifepoints changed to float

+++ Penteract/Assets/Scripts/Character.h
@@ -14,16 +14,15 @@ class Character
 public:
 	Character() {}
 
-	virtual void Hit(int damage_);
+	virtual void Hit(float damage_);
 	void Recover(int recoveryLife_);
 
 	void SetTotalLifePoints(int totalLifePoints_);
 	void SetDamageHit(int damageHit_);
-
 public:
 	bool isAlive = true;
-	int lifePoints = 1;
-	int damageHit = 1;
+	float lifePoints = 1;
+	float damageHit = 1;
 	float movementSpeed = 1.0f;
 	ComponentAgent* agent = nullptr;
 	ComponentAnimation* compAnimation = nullptr;
@@ -31,5 +30,5 @@ public:
 	GameObject* characterGameObject = nullptr;
 
 private:
-	int totalLifePoints = 1;
+	float totalLifePoints = 1;
 };
